Changes implementation of locale setting from d8a8a07
The way it was done in d8a8a07 is not correct as the
gettext would only work for the strings in the main
function.
gettext should be called at the top of the module.
And this is what this commit does.
Some explanations:
In addition to above gettext  autodetects locale
settings and finds correct files with translation
- there is no need to do it manually.
That also means that whatever the get_current_locale
function returns is not related in any way to gettext
autodetection mechanism. Both will usually be consistent
but this is not guaranteed.
get_current_locale will be used for LibreOffice preferred
Language pack and will have nothing to do with localization
of the lomanager2 itself.