Adds QAbstractTableModel based interface
This is the Model-View/Controller scheme. The interface will be used
by the QUI (View) to interact with the data model (not yet implemented)
It will allow the GUI to react to any data changes and
automaticaly refresh the View.
Not yet ready data model (independant of interface above) will
in turn encapsulate entire package selection logic.