Moves Java install logic to MainLogic, changes what is passed to subprocedure...
This begins a transition to a new logic where packege name resolution (to actual rpm names) will be
done in the subprocedures and not the MainLogic.
Subprocedures are going to receive the list of virtual packages.
Each VP will be holding information needed by install/remove functions.
This requires changes in the subproceduress which will follow in next commits