Fixes worker thread implementation...
...Now thread is created in one of Adapter's slot methods on demand (not in the __init__).
This allows sending up to date parameters set by the user in the GUI.
The thread worker now runs only a single method of the MainLogic to apply changes but passes on the
installation mode as a variable (this depends on user selection in the GUI).
Underlying MainLogic method is not yet modified to support this.