Replaces emitting status with adding msg for user and returning
Instead of emitting status signal and returning from the method add a message for the user
to the list of messages and return. List off messages is being checked anyway after a thread finishes.
That way only one method in the adapter is responsible for controlling the information dialog
(excluding progress dialog)