Removes old method of setting mark state by passing strings
Since EditRole is now defined in model's data() method (and it returns boolen values for install/remove
fields as expected) the default delagate created by Qt is a combobox (not QLineEdit used because only
DisplayRole existed and it always returns str) returning True/False.
Also the new custom delegate can return bool (not str) so this is no longer needed.