Fixes not being able to force specific Clipart version using apt-get
When there is a network connection during install with 'apt-get install' apt-get will check repo server
by itself for the latest version of rpm-s it was given to install. If it finds a newer version
it will download it and install it instead of the rpm passed as a parameter.
To remedy this problem Clipart is now installed the same way as LibreOffice - using 'rpm -Uvh'