Fixes delegate button size definition
sizeHint method was added to delegate to achieve this.
Now created button size is defined as a size, evaluated at run time, of a fixed text string. So as long
as the column title is shorter, this text will define button size.
Delegate biding to view was moved from adapter to view definition itself.